requirements.txt
fastapi==0.75.2
uvicorn[standard]==0.23.2
jinja2==3.0.3
python-dotenv==1.0.1
elasticsearch[async]==7.17.12
pymongo==4.7.3
minio==7.2.7
psycopg2-binary
